Temp extract
Added some lines so you can get the temperature from the sensors, we should be able to get the pressure as well but wasn't able to do that. I also switched it so the time vector is pulled from the sensor rather than calculating it with the sampling rate. Tried to output sampling rate but couldn't get it to work... would be helpful to know which sensor all the data came from as well for outputs.